Fix left-over bug from previous "unique index" bug
Fixes up a bug that was left over from https://github.com/hdrake/regionate/commit/68e70178ef373b384317a80044d1d99008fd107b.
The indices from contourpy.contour_generator were already "looped", meaning that the origin and endpoint were the same, so
that the polygon is closed. Before the prior commit, we were removing any duplicate points (assuming they were due to some
unspecified error) and then only adding back in the duplicated endpoint/origin to "loop" the boundary points.
However, this ignored the fact that intermediate points can also be duplicated for valid reasons, as shown in
https://github.com/hdrake/regionate/issues/4#issue-1816504475.

When we removed the "unique_list" call in the prior commit, we left the original looping in addition to the final looping,
which resulting in erroneous double-counting of transports. This is fixed in this commit. We also re-run all of the
example notebooks and confirm that the divergence theorem is still satisfied for all of the different regions.