Handle weird corner cases for wrap-around masks
For some reason, I found that for wrap-around masks (i.e. with boundaries
that are period in longitude), there were examples where the `x` index
in the mask was erroneously all True. I think it had to do with how I was
being clever about making the mask go all the way to Antarctica and back.

I think this addresses the issue.