Finishes implementation of `BoundedRegion` class, including children/parents
In particular, it now provides the child sections with not just a pointer
back to the parent gridded section, but also the corresponding indices of
both the corner and velocity points of the parent section. This means that
one can very easily diagnose transports across child sections from a
single calculation for the parent section!

Also changes convention for `lons_c` and `lats_c`, making them have the same
length as `i_c` and `j_c`, and making it clear that they correspond to
corner points as well.